Dropdown moved into its own file in components
